Added log normal fire size fitting
Added a lognormal distribution fit to grab the fire size and size sd
inputs required by the DFFS. Because the parameters (mu and s) are
supposed to estimate the location of the mean fire size, given the
lognormal distribution, I both calculated the true mean of the log of
fire size as well as the mean of the fitted distribution to the data.
Not the same obviously.. but not too far off either.