Modified cleanRaws() to include a windfix flag
Allows the reversal of wind directions based on a flag: if 'yes', switch
the wind azimuths. Useful given that in the case of the severe weather,
you don't want to be unreversing the azimuths if you end up running the
same met data through the RAWS cleaning process more than once.