Created a stand map generation function
Quickly built a function to read in an existing LANDIS input raster, in
this case the initial communities map, and use it to create a raster
mask (delineate inactive cells) as well as to output a geotiff stand
map, where each stand is represented by a unique integer  across the
landscape. Currently, stand size is equal to IC grid cell size.