Redesigned the first  cut of stand map gen
Because of the resampling required of the management unit raster, the
stand map occasionally would show the same stand across 2 management
units. Not sure if that was really happening however... the data type of
the stand map was set to Int16 -- and given the number cells in the
raster (277 * 335), there weren't enough unique values to populate the
raster (max at 2^16). Set the data type to int_32, should be sufficient.